# server/swagger.yml
openapi: "3.0.3"
info:
  title: SugarBox API by arairon
  version: 0.1.0
  description: A simple storage for SugarCube saves.
  license:
    name: GPL-3.0
    url: https://opensource.org/license/gpl-3-0
  contact:
    name: arai.icu
    url: https://arai.icu
    email: arai@arai.icu
servers:
  - url: https://sugarbox.arai.icu/api
  #- url: http://127.0.0.1:3000/api
paths:
  /auth/register:
    post:
      summary: Register a new user
      description: Registers a new user with a username, email, and password.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegisterSchema"
      responses:
        "201":
          description: Successfully registered user
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      user:
                        $ref: "#/components/schemas/UserReturnSchema"
                      session:
                        type: object
        "400":
          description: Validation error or username/email already taken
        "500":
          description: Internal server error
  /auth/login:
    post:
      summary: Log in a user
      description: Authenticates a user with their username/email and password.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserLoginSchema"
      responses:
        "200":
          description: Successfully logged in
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      user:
                        $ref: "#/components/schemas/UserReturnSchema"
                      session:
                        type: object
        "400":
          description: Invalid credentials or validation error
        "404":
          description: User not found
        "500":
          description: Internal server error
  /auth/logout:
    post:
      summary: Log out a user
      description: Logs out the currently authenticated user by invalidating their session.
      parameters:
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
          description: Bearer token in the format "Bearer ACCESSTOKEN"
      responses:
        "200":
          description: Successfully logged out
        "401":
          description: Unauthorized or invalid token
        "500":
          description: Internal server error
  /auth/refresh:
    get:
      summary: Refresh session token
      description: Refreshes the access and refresh tokens for the current session.
      parameters:
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
          description: Bearer token in the format "Bearer REFRESHTOKEN"
      responses:
        "200":
          description: Successfully refreshed session
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      user:
                        $ref: "#/components/schemas/UserReturnSchema"
                      session:
                        type: object
        "403":
          description: Invalid or expired refresh token
        "404":
          description: User not found
        "500":
          description: Internal server error
  /auth/self:
    get:
      summary: Get current user information
      description: Retrieves information about the currently authenticated user.
      parameters:
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
          description: Bearer token in the format "Bearer ACCESSTOKEN"
      responses:
        "200":
          description: Successfully retrieved user information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserReturnSchema"
        "401":
          description: Unauthorized or invalid token
        "404":
          description: User not found
        "500":
          description: Internal server error
  /auth:
    get:
      summary: Get user and role from token
      description: Decodes the provided token to retrieve the user ID and role.
      parameters:
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
          description: Bearer token in the format "Bearer ACCESSTOKEN"
      responses:
        "200":
          description: Successfully retrieved user and role
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    type: integer
                  role:
                    type: string
        "401":
          description: Unauthorized or invalid token
...
  /user/self:
    patch:
      summary: Update user information
      description: Updates the authenticated user's account information, such as username, email, or password.
      parameters:
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
          description: Bearer token in the format "Bearer ACCESSTOKEN"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserPatchSchema"
      responses:
        "200":
          description: Successfully updated user information
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
                  data:
                    type: object
        "400":
          description: Validation error or username/email already taken
        "403":
          description: Invalid authentication data
        "500":
          description: Internal server error
...
components:
  schemas:
    UserRegisterSchema:
      type: object
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 32
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 3
          maxLength: 256
    UserLoginSchema:
      type: object
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 32
        password:
          type: string
          minLength: 3
          maxLength: 256
    UserReturnSchema:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        displayname:
          type: string
        email:
          type: string
        emailConfirmed:
          type: boolean
        role:
          type: string
    UserPatchSchema:
      type: object
      properties:
        displayname:
          type: string
          description: Display name of the user.
          minLength: 3
          maxLength: 64
        username:
          type: string
          description: Username of the user.
          minLength: 3
          maxLength: 32
        email:
          type: string
          description: Email address of the user.
          format: email
        password:
          type: string
          description: Password of the user.
          minLength: 3
          maxLength: 256
...